Modify type definitions
